[Bugfix #584] Refactor: Extract writeMessageToSession to shared module
Move pacing logic to message-write.ts to avoid circular dependency
between tower-routes.ts and tower-cron.ts. Also fix tower-cron.ts
cron message delivery to use the same pacing.
